refactored code
 - Moved viewholder into adapter
 - Moved adapter and fragment into subpackage
