--- app/src/main/java/com/karthikb351/vitinfo2/LayoutManagers/CourseLayoutManager.java
@@ -1,43 +1,0 @@
-/*
- * VITacademics
- * Copyright (C) 2015  Aneesh Neelam <neelam.aneesh@gmail.com>
- * Copyright (C) 2015  Gaurav Agerwala <gauravagerwala@gmail.com>
- * Copyright (C) 2015  Pulkit Juneja <pulkit.16296@gmail.com>
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-package com.karthikb351.vitinfo2.LayoutManagers;
-
-import android.support.v7.widget.RecyclerView;
-
-/**
- * Created by gaurav on 15/6/15.
- */
-public class CourseLayoutManager extends RecyclerView.LayoutManager {
-
-    @Override
-    public RecyclerView.LayoutParams generateDefaultLayoutParams() {
-        return null;
-    }
-    @Override
-    public void scrollToPosition(int position) {
-
-    }
-
-    @Override
-    public void onLayoutChildren(RecyclerView.Recycler recycler, RecyclerView.State state) {
-        super.onLayoutChildren(recycler, state);
-    }
-}
